requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1
PyYAML>=6.0
matplotlib>=3.5.0
